Adding Hyperlinked Cross-References {3.1.Xref} etc.
Actually, this _restores_ the hyperlinked cross-references from the Word original.  Now, in the rendered document, each cross-reference within the document is a hyperlink to that section. 
Currently the sections are labelled by their outline numbers, so the links reference the section number ( href="#3.2.Sec").  In future, we will give semantic names to sections (Conf.,  Misc., etc.) and then they will look like href="#Conf.Engage.Sec", etc.  This is a pretty robust way of building a document and having the cross-references work by themselves.  Even if you reorganize the order of sections or delete some, the semantic cross-reference only needs to be changed in one place - at the Xref widget.